+++ src/SoapCore.Tests/MessageInspector/MessageInspectorMock.cs
@@ -1,6 +1,7 @@
 using System.IO;
 using System.ServiceModel.Channels;
 using System.Xml;
+using SoapCore.Extensibility;
 
 namespace SoapCore.Tests.MessageInspector
 {
