+++ src/SoapCore.Tests/MessageInspectors/MessageInspector/MessageInspectorMock.cs
@@ -1,5 +1,6 @@
 using System.ServiceModel;
 using System.ServiceModel.Channels;
+using SoapCore.Extensibility;
 
 namespace SoapCore.Tests.MessageInspectors.MessageInspector
 {
