+++ src/SoapCore.Tests/TrailingServicePathTuner/MockModelBounder.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Reflection;
 using System.Text;
+using SoapCore.Extensibility;
 
 namespace SoapCore.Tests
 {
