+++ src/SoapCore/DefaultFaultExceptionTransformer.cs
@@ -1,5 +1,6 @@
 using System;
 using System.ServiceModel.Channels;
+using SoapCore.Extensibility;
 
 namespace SoapCore
 {
