+++ src/SoapCore/Extensibility/IOperationInvoker.cs
@@ -1,11 +1,7 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Reflection;
-using System.Text;
 using System.Threading.Tasks;
 
-namespace SoapCore
+namespace SoapCore.Extensibility
 {
 	public interface IOperationInvoker
 	{
